requirement.txt
streamlit==1.18.0
tensorflow==2.13.0
pillow==9.3.0
numpy==1.23.5
gdown==4.7.0
boto3==1.26.0
